+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 ### Added 
 - ae_event_t to wrap eventfd
 - ae_pool_reset()
+- default handler for missing options
 ### Fixed
 - io.h was missing from the install target in Makefile.am
 - fixed count in ae_mux_rm